+++ crates/circuit/src/packed_instruction.rs
@@ -379,11 +379,11 @@ impl Operation for PackedOperation {
         }
     }
     #[inline]
-    fn num_qubits(&self) -> u32 {
+    fn num_qubits(&self) -> usize {
         self.view().num_qubits()
     }
     #[inline]
-    fn num_clbits(&self) -> u32 {
+    fn num_clbits(&self) -> usize {
         self.view().num_clbits()
     }
     #[inline]
